\name{atmopt-package}
\alias{atmopt-package}
\alias{atmopt}
\docType{package}
\title{Analysis-of-marginal-Tail-Means}
\description{
The 'atmopt' package provides functions for implementing the ATM optimization method in Mak and Wu (2018+) <arXiv:1712.03589>.
}
\details{
\tabular{ll}{
Package: \tab atmopt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-10-19\cr
License: \tab GPL (>= 2)\cr
}
The 'atmopt' package provides functions for implementing the Analysis-of-marginal-Tail-Means (ATM) method in Mak and Wu (2018+). ATM is a robust method for discrete, black-box optimization, where function evaluations are expensive and limited.}
\author{
Simon Mak

Maintainer: Simon Mak <smak6@gatech.edu>
}
\references{
Mak, S. and Wu, C. F. J. (2018+). Analysis-of-marginal-Tail-Means (ATM): a robust method for discrete black-box optimization. Under review.
}
\keyword{ Robust optimization, discrete optimization, black-box optimization }
